Adapt for MultiBUGS
This diff is mostly just a find-and-replace of "MultiBUGS" for "OpenBUGS"

However, it also adds n.workers argument to bugs, so that the number of workers can be set. The default value is obtained using parallel::detectCores()

DIC is now obtained using dicSetS() rather than dicSet(), which does not work in MultiBUGS.

It also warns that seeds can't be set in MultiBUGS (yet).