Monitor all components of outlier in Stacks
R2MultiBUGS always sets summary monitors, but it has never been
possible to set a summary monitor on a single component of a
parameter. In OpenBUGS, trying to set such a summary monitor in a
script yields an error, but the script keeps going, whereas MultiBUGS
is more conservative.

To avoid this, we now monitor all components.